Disable garbage collection for calculating PBKDF2
We previously attempted to avoid garbage collection while benchmarking
by triggering a garbage collection before the loop. This would provide
an unrepresentative sample, though, because in real usage garbage
collection would *not* be triggered ahead of time.

Better is to disable garbage collection entirely (rather than trying to
"fake" it). And best is to do this step every time the KDF is run, so
that the benchmarked case is representative of actual use.